Fix error: Please use a locale setting which supports utf-8 (#260)
* Fix error: Please use a locale setting which supports utf-8

The yocto master print's out that error while using `bitbake`:

```
Please use a locale setting which supports utf-8.
Python can't change the filesystem locale after loading so we need a
utf-8 when python starts or things won't work.
```

* Wrong config file. ~/.profile wasn't working

* Inject the locales via ENV

So that they are not only available to the user `build`, but also to all other users e.g. `jenkins`.